added JwtSignup functionality
