--- api-gateway/src/main/java/com/example/apigateway/config/AuthenticationFilter.java
@@ -1,33 +1,0 @@
-package com.example.apigateway.config;
-
-import com.netflix.zuul.ZuulFilter;
-import com.netflix.zuul.context.RequestContext;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-
-@Component
-public class AuthenticationFilter extends ZuulFilter {
-    @Override
-    public String filterType() {
-        return null;
-    }
-
-    @Override
-    public int filterOrder() {
-        return 0;
-    }
-
-    @Override
-    public boolean shouldFilter() {
-        return false;
-    }
-
-    @Override
-    public Object run() {
-        RequestContext ctx = RequestContext.getCurrentContext();
-        String un = SecurityContextHolder.getContext().getAuthentication().getName();
-        ctx.addZuulRequestHeader("username", un);
-        return null;
-    }
-
-}
