--- api-gateway/src/main/java/com/example/apigateway/repository/UserRepository.java
@@ -1,27 +1,0 @@
-package com.example.apigateway.repository;
-
-import com.example.apigateway.bean.UserBean;
-import java.sql.ResultSet;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class UserRepository {
-
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
-
-    public UserBean getUserByUsername(String username){
-        String sql = "SELECT * FROM users WHERE username = ?";
-
-        return jdbcTemplate.queryForObject(sql, new Object[]{username}, (rs, rowNum) ->
-                new UserBean(
-                        rs.getLong("id"),
-                        rs.getString("email"),
-                        rs.getString("username"),
-                        rs.getString("password")
-                ));
-    }
-
-}
