+++ api-gateway/src/main/resources/application-local.properties
@@ -1,0 +1,15 @@
+spring.application.name=api-gateway
+
+management.endpoints.web.expose=*
+
+# Add color to log output
+spring.output.ansi.enabled=ALWAYS
+
+eureka.client.serviceUrl.defaultZone=http://localhost:8761/eureka/
+
+spring.datasource.url=jdbc:postgresql://localhost:5432/pgdev
+#spring.datasource.username=postgres
+#spring.datasource.password=password
+
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults=false
