+++ eureka-server/pom.xml
@@ -33,7 +33,7 @@
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-netflix-eureka-server</artifactId>
-			<version>2.1.2.RELEASE</version>
+			<version>2.1.3.RELEASE</version>
 		</dependency>
 
 		<dependency>
