+++ users-api/src/main/java/com/example/usersapi/service/UserService.java
@@ -1,5 +1,6 @@
 package com.example.usersapi.service;
 
+import com.example.usersapi.model.JwtResponse;
 import com.example.usersapi.model.User;
 import org.springframework.http.HttpStatus;
 
@@ -13,7 +14,7 @@ public interface UserService {
 
     public HttpStatus deleteUser(long id);
 
-    public HttpStatus createUser(User user);
+    public JwtResponse createUser(User user);
 
     public HttpStatus updateUser(long id, User userRequest);
 }
