+++ users-api/src/main/resources/application-local.properties
@@ -1,0 +1,16 @@
+server.port=8081
+
+spring.application.name=users
+
+eureka.client.serviceUrl.defaultZone=http://localhost:8761/eureka/
+
+# Add color to log output
+spring.output.ansi.enabled=ALWAYS
+
+spring.datasource.url=jdbc:postgresql://localhost:5432/pgdev
+#spring.datasource.username=postgres
+#spring.datasource.password=password
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+
+spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults=false
