+++ users-api/src/main/resources/application.properties
@@ -1,2 +1,2 @@
-spring.profiles.active=dev
+spring.profiles.active=local
 jwt.secret=pancakes
