+++ users-api/src/main/resources/db/migration/V1__create_users.sql
@@ -1,6 +1,6 @@
 CREATE TABLE users (
     id SERIAL,
-    email VARCHAR(100) NOT NULL,
-    password VARCHAR(50),
-    username VARCHAR(50)
+    email VARCHAR(100) UNIQUE NOT NULL,
+    password VARCHAR,
+    username VARCHAR(50) UNIQUE NOT NULL
 );
