+++ users-api/src/main/resources/db/migration/V2__create_posts.sql
@@ -1,6 +1,6 @@
 CREATE TABLE posts (
     id SERIAL,
-    title VARCHAR(200),
+    title VARCHAR(200) NOT NULL,
     description VARCHAR(200),
-    user_id INT
+    user_id INT NOT NULL
 );
