+++ users-api/src/main/resources/db/migration/V3__create_comments.sql
@@ -1,6 +1,6 @@
 CREATE TABLE comments (
     id SERIAL,
     text VARCHAR(200),
-    post_id INT,
-    user_id INT
+    post_id INT NOT NULL,
+    user_id INT NOT NULL
 );
